Change f-string to prinf-style log formatting
