+++ src/app/api/v1/utils.py
@@ -16,23 +16,3 @@ def get_body(request_model: Type[BaseModel]):
         abort(422, description=error_message)
     else:
         return body
-
-
-def before_request_log():
-    logger.info("{}: {}", request.method, request.path)
-    logger.debug("{}", request.args)
-    if (
-        request.content_type == "application/json"
-        and int(request.headers.get("Content-Length")) > 0
-    ):
-        logger.debug("{}", request.get_json())
-
-
-def after_request_log(response):
-    logger.debug(
-        "Response for {} {}: {}",
-        request.method,
-        request.path,
-        response.get_json(),
-    )
-    return response
